+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Fixed
+- continuations breaking after escaped comments
+- escape signs sometimes left in the output
+### Added
+- escaping comment signs
+- quoting parameter names
+- multi-line quoting
 
 ## [0.2.2] - 2025-04-02
 ### Fixed